refactor: swap transaction status
- remove setApproval & setSwap
- fix updateSwapSteps
- change the GenericButton to SwapModalButton
