+++ components/02-molecules/index.tsx
@@ -4,5 +4,6 @@ export * from "./FilterOffers";
 export * from "./NftCard";
 export * from "./NftsList";
 export * from "./OfferSummary";
+export * from "./ProgressStatus";
 export * from "./SwapStation";
 export * from "./TheHeader";
